Redo how the multi gradient editor works (again!)
Also make it do linear conversion in the updateTexture method instead of changing the import on the texture, this makes it so realtime actually shows the correct thing.
Remove the extention from the file saving prompt so it's more user friendly.
Made it so you can drag the gradients when you've enabled the Reorder button.